% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baap.R
\docType{package}
\name{baap}
\alias{baap}
\title{baap: A Companion of the book "Basic and Advanced Psychometrics in R"}
\description{
The baap package provides necessary data frames used through out the book and some neat functions
}
\section{baap Data-frames}{

\enumerate{
\item Rotter: Psychometric validation data of Bangla Rotter's Internal- External Data
\item Gantt and Spot: Two sample data-frames for creating project management Gantt chart
\item FFMQ.CFA: Structural Validitation data of Bangla Five Factor Mindfulness Questionnaire.
\item FFMQ.Val:Correlational Validity evidences of Bangla Five Factor Mindfulness Questionnaire.
}
}

\section{baap functions}{

This packages includes some neat functions. Some of the functions presented here are from
stack overflow and other resources that we find very helpful in formatting the outputs
in a publication ready state.
\enumerate{
\item fac.table(): Creates a publication ready table from the output of psych based factor analysis.
\item des.table(): Creates a publication ready descriptive table of Item analysis with the reporting of normality assumptions.
\item normality.loop(): Compute normality test on the whole data frame. No grouping required.
\item bedTime(): Calculate total time spent in bed from the sleep log entry.
}

It is advisable that you load \strong{tidyverse} along with \strong{baap}
}

