% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des.tab.R
\name{des.tab}
\alias{des.tab}
\title{A Function for Descriptive data for item analysis.}
\usage{
des.tab(df, reverse = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{reverse}{If TRUE, will provide indicate which items had a negative correlation and reverse them}
}
\value{
Returns a summary table of descriptives in a data frame structure.
}
\description{
This function will create a publication ready essential descriptive table for item analysis.
Normality is tested using shapiro.test from base stats with Bonferroni Correction.
}
\examples{
data <- tabledown::Rotter[, 11:31]
table <- des.tab(data)
}
