% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FFMQ.CFA}
\alias{FFMQ.CFA}
\title{Structural Validity data of FFMQ}
\format{
A data frame with 277 rows and 47 variables:
\describe{
\item{\code{ID}}{double COLUMN_DESCRIPTION}
\item{\code{Gender}}{character COLUMN_DESCRIPTION}
\item{\code{Education}}{character COLUMN_DESCRIPTION}
\item{\code{Education Years}}{double COLUMN_DESCRIPTION}
\item{\code{Income}}{double COLUMN_DESCRIPTION}
\item{\code{Profession}}{character COLUMN_DESCRIPTION}
\item{\code{Marital Status}}{character COLUMN_DESCRIPTION}
\item{\code{Social_status}}{double COLUMN_DESCRIPTION}
\item{\code{item1}}{double COLUMN_DESCRIPTION}
\item{\code{item2}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem3}}{double COLUMN_DESCRIPTION}
\item{\code{item4}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem5}}{double COLUMN_DESCRIPTION}
\item{\code{item6}}{double COLUMN_DESCRIPTION}
\item{\code{item7}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem8}}{double COLUMN_DESCRIPTION}
\item{\code{item9}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem10}}{double COLUMN_DESCRIPTION}
\item{\code{item11}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem12}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem13}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem14}}{double COLUMN_DESCRIPTION}
\item{\code{item15}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem16}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem17}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem18}}{double COLUMN_DESCRIPTION}
\item{\code{item19}}{double COLUMN_DESCRIPTION}
\item{\code{item20}}{double COLUMN_DESCRIPTION}
\item{\code{item21}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem22}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem23}}{double COLUMN_DESCRIPTION}
\item{\code{item24}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem25}}{double COLUMN_DESCRIPTION}
\item{\code{item26}}{double COLUMN_DESCRIPTION}
\item{\code{item27}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem28}}{double COLUMN_DESCRIPTION}
\item{\code{item29}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem30}}{double COLUMN_DESCRIPTION}
\item{\code{item31}}{double COLUMN_DESCRIPTION}
\item{\code{item32}}{double COLUMN_DESCRIPTION}
\item{\code{item33}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem34}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem35}}{double COLUMN_DESCRIPTION}
\item{\code{item36}}{double COLUMN_DESCRIPTION}
\item{\code{item37}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem38}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem39}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{https://github.com/masiraji/tabledown/tree/main/data-raw}
}
\usage{
FFMQ.CFA
}
\description{
This is the structural validation data of Bangla Five Facet Mindfulness Questionnaire
}
\keyword{datasets}
