% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FFMQ.Val}
\alias{FFMQ.Val}
\title{Correlational based Valididity evidence of FFMQ}
\format{
A data frame with 255 rows and 106 variables:
\describe{
\item{\code{id}}{double COLUMN_DESCRIPTION}
\item{\code{Age}}{double COLUMN_DESCRIPTION}
\item{\code{Gender}}{double COLUMN_DESCRIPTION}
\item{\code{Education Years}}{double COLUMN_DESCRIPTION}
\item{\code{Profession}}{character COLUMN_DESCRIPTION}
\item{\code{Marital Status}}{character COLUMN_DESCRIPTION}
\item{\code{Social_Status}}{double COLUMN_DESCRIPTION}
\item{\code{item1}}{double COLUMN_DESCRIPTION}
\item{\code{item2}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem3}}{double COLUMN_DESCRIPTION}
\item{\code{item4}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem5}}{double COLUMN_DESCRIPTION}
\item{\code{item6}}{double COLUMN_DESCRIPTION}
\item{\code{item7}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem8}}{double COLUMN_DESCRIPTION}
\item{\code{item9}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem10}}{double COLUMN_DESCRIPTION}
\item{\code{item11}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem12}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem13}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem14}}{double COLUMN_DESCRIPTION}
\item{\code{item15}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem16}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem17}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem18}}{double COLUMN_DESCRIPTION}
\item{\code{item19}}{double COLUMN_DESCRIPTION}
\item{\code{item20}}{double COLUMN_DESCRIPTION}
\item{\code{item21}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem22}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem23}}{double COLUMN_DESCRIPTION}
\item{\code{item24}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem25}}{double COLUMN_DESCRIPTION}
\item{\code{item26}}{double COLUMN_DESCRIPTION}
\item{\code{item27}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem28}}{double COLUMN_DESCRIPTION}
\item{\code{item29}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem30}}{double COLUMN_DESCRIPTION}
\item{\code{item31}}{double COLUMN_DESCRIPTION}
\item{\code{item32}}{double COLUMN_DESCRIPTION}
\item{\code{item33}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem34}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem35}}{double COLUMN_DESCRIPTION}
\item{\code{item36}}{double COLUMN_DESCRIPTION}
\item{\code{item37}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem38}}{double COLUMN_DESCRIPTION}
\item{\code{Ritem39}}{double COLUMN_DESCRIPTION}
\item{\code{EI1}}{character COLUMN_DESCRIPTION}
\item{\code{EI2}}{character COLUMN_DESCRIPTION}
\item{\code{EI3}}{character COLUMN_DESCRIPTION}
\item{\code{EI4}}{character COLUMN_DESCRIPTION}
\item{\code{EI5}}{character COLUMN_DESCRIPTION}
\item{\code{EI6}}{character COLUMN_DESCRIPTION}
\item{\code{EI7}}{character COLUMN_DESCRIPTION}
\item{\code{EI8}}{character COLUMN_DESCRIPTION}
\item{\code{EI9}}{character COLUMN_DESCRIPTION}
\item{\code{EI10}}{character COLUMN_DESCRIPTION}
\item{\code{EI11}}{character COLUMN_DESCRIPTION}
\item{\code{EI12}}{character COLUMN_DESCRIPTION}
\item{\code{EI13}}{character COLUMN_DESCRIPTION}
\item{\code{EI14}}{character COLUMN_DESCRIPTION}
\item{\code{EI15}}{character COLUMN_DESCRIPTION}
\item{\code{EI16}}{character COLUMN_DESCRIPTION}
\item{\code{EI17}}{character COLUMN_DESCRIPTION}
\item{\code{EI18}}{character COLUMN_DESCRIPTION}
\item{\code{EI19}}{character COLUMN_DESCRIPTION}
\item{\code{EI20}}{character COLUMN_DESCRIPTION}
\item{\code{EI21}}{character COLUMN_DESCRIPTION}
\item{\code{EI22}}{character COLUMN_DESCRIPTION}
\item{\code{EI23}}{character COLUMN_DESCRIPTION}
\item{\code{EI24}}{character COLUMN_DESCRIPTION}
\item{\code{EI25}}{character COLUMN_DESCRIPTION}
\item{\code{EI26}}{character COLUMN_DESCRIPTION}
\item{\code{EI27}}{character COLUMN_DESCRIPTION}
\item{\code{EI28}}{character COLUMN_DESCRIPTION}
\item{\code{EI29}}{character COLUMN_DESCRIPTION}
\item{\code{EI30}}{character COLUMN_DESCRIPTION}
\item{\code{EI31}}{character COLUMN_DESCRIPTION}
\item{\code{EI32}}{character COLUMN_DESCRIPTION}
\item{\code{EI33}}{character COLUMN_DESCRIPTION}
\item{\code{EI34}}{character COLUMN_DESCRIPTION}
\item{\code{O1}}{character COLUMN_DESCRIPTION}
\item{\code{O2}}{character COLUMN_DESCRIPTION}
\item{\code{O3}}{character COLUMN_DESCRIPTION}
\item{\code{O4}}{character COLUMN_DESCRIPTION}
\item{\code{O5}}{character COLUMN_DESCRIPTION}
\item{\code{O6}}{character COLUMN_DESCRIPTION}
\item{\code{O7}}{character COLUMN_DESCRIPTION}
\item{\code{O8}}{character COLUMN_DESCRIPTION}
\item{\code{O9}}{character COLUMN_DESCRIPTION}
\item{\code{O10}}{character COLUMN_DESCRIPTION}
\item{\code{E1}}{character COLUMN_DESCRIPTION}
\item{\code{E2}}{character COLUMN_DESCRIPTION}
\item{\code{E3}}{character COLUMN_DESCRIPTION}
\item{\code{E4}}{character COLUMN_DESCRIPTION}
\item{\code{E5}}{character COLUMN_DESCRIPTION}
\item{\code{E6}}{character COLUMN_DESCRIPTION}
\item{\code{E7}}{character COLUMN_DESCRIPTION}
\item{\code{E8}}{character COLUMN_DESCRIPTION}
\item{\code{N1}}{character COLUMN_DESCRIPTION}
\item{\code{N2}}{character COLUMN_DESCRIPTION}
\item{\code{N3}}{character COLUMN_DESCRIPTION}
\item{\code{N4}}{character COLUMN_DESCRIPTION}
\item{\code{N5}}{character COLUMN_DESCRIPTION}
\item{\code{N6}}{character COLUMN_DESCRIPTION}
\item{\code{N7}}{character COLUMN_DESCRIPTION}
\item{\code{N8}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
FFMQ.Val
}
\description{
Correlational based Valididity evidence of Bangla FFMQ
}
\keyword{datasets}
