% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gantt}
\alias{Gantt}
\title{Gantt Data}
\format{
A data frame with 25 rows and 4 variables:
\describe{
\item{\code{wp}}{character Main Component}
\item{\code{activity}}{character Activities}
\item{\code{start_date}}{character Start Date}
\item{\code{end_date}}{character End Date}
}
}
\source{
\url{https://github.com/masiraji/tabledown/tree/main/data-raw}
}
\usage{
Gantt
}
\description{
Demo project breakdown to create Gantt
}
\keyword{datasets}
