% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggreliability_plotly.R
\name{ggreliability_plotly}
\alias{ggreliability_plotly}
\title{A Function for Creating   Item Response Theory based reliability plot based on plotly.}
\usage{
ggreliability_plotly(dataframe, model)
}
\arguments{
\item{dataframe}{your data.}

\item{model}{A mirt package fitted object.}
}
\value{
A publication quality reliability plot (dashed line). Output object is a ggplot object.
}
\description{
This function will create  Item Response Theory based based reliability plot with standard error using ggplot2 and plotly from objects created from
mirt pack.
Using ggplot2 will enable the user to modify the Item plot.
}
\examples{
data <- tabledown::Rotter[, 11:31]
model <- mirt::mirt(data, model = 1, itemtype = '2PL')

plot <- ggreliability_plotly(data, model)
}
