% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtestinfo_se.R
\name{ggtestinfo_se}
\alias{ggtestinfo_se}
\title{A Function for Creating  Publication Quality IRT test information plot with standard error.}
\usage{
ggtestinfo_se(dataframe, model)
}
\arguments{
\item{dataframe}{your data.}

\item{model}{A mirt package fitted object.}
}
\value{
A publication quality Test information plot with standard error (dashed line). Output object is a ggplot object.
}
\description{
This function will create publication worthy IRT Test information plot with standard error using ggplot2 from objects created from
mirt pack.
Using ggplot2 will enable the user to modify the Item plot.
}
\examples{
data <- tabledown::Rotter[, 11:31]
model <- mirt::mirt(data, model = 1, itemtype = '2PL')

plot <- ggtestinfo(data, model)
}
