% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gttab.R
\name{gt_tab}
\alias{gt_tab}
\title{A Function for gtExtra package friendly data summary.}
\usage{
gt_tab(dataframe, recode_code)
}
\arguments{
\item{recode_code}{Recode key}

\item{datafrmae}{Dataframe with all items.}
}
\value{
A publication ready descriptive summary table in png format.
}
\description{
This function will gtExtra package friendly data summary using the datafrmae provided
psych pack.
}
\examples{
data <- tabledown::FFMQ.CFA[, 9:18]
recode_code <- c( "1" = "Never or very rarely true", "2" = "Rarely true", "3"= "Sometimes true","4" = "Often true","5" = "Very often or always true")
sample_tab <- gt_tab(gt.data,recode_code)
}
