% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rotter}
\alias{Rotter}
\title{Validation Data of Bangla Rotter I-E Scale}
\format{
A data frame with 478 rows and 91 variables:
\describe{
\item{\code{id}}{double Id}
\item{\code{sample}}{character EFA or CEA}
\item{\code{Age}}{double Age}
\item{\code{Gender}}{character Gender}
\item{\code{Educational Status}}{character Educational Status}
\item{\code{Education Years}}{double COLUMN_DESCRIPTION}
\item{\code{Income}}{double COLUMN_DESCRIPTION}
\item{\code{Religion}}{double COLUMN_DESCRIPTION}
\item{\code{Marital Status}}{double COLUMN_DESCRIPTION}
\item{\code{Social Stance}}{double COLUMN_DESCRIPTION}
\item{\code{item2}}{double COLUMN_DESCRIPTION}
\item{\code{item3}}{double COLUMN_DESCRIPTION}
\item{\code{item4}}{double COLUMN_DESCRIPTION}
\item{\code{item5}}{double COLUMN_DESCRIPTION}
\item{\code{item6}}{double COLUMN_DESCRIPTION}
\item{\code{item7}}{double COLUMN_DESCRIPTION}
\item{\code{item9}}{double COLUMN_DESCRIPTION}
\item{\code{item10}}{double COLUMN_DESCRIPTION}
\item{\code{item11}}{double COLUMN_DESCRIPTION}
\item{\code{item12}}{double COLUMN_DESCRIPTION}
\item{\code{item13}}{double COLUMN_DESCRIPTION}
\item{\code{item15}}{double COLUMN_DESCRIPTION}
\item{\code{item16}}{double COLUMN_DESCRIPTION}
\item{\code{item17}}{double COLUMN_DESCRIPTION}
\item{\code{item18}}{double COLUMN_DESCRIPTION}
\item{\code{item20}}{double COLUMN_DESCRIPTION}
\item{\code{item21}}{double COLUMN_DESCRIPTION}
\item{\code{item22}}{double COLUMN_DESCRIPTION}
\item{\code{item23}}{double COLUMN_DESCRIPTION}
\item{\code{item25}}{double COLUMN_DESCRIPTION}
\item{\code{item26}}{double COLUMN_DESCRIPTION}
\item{\code{item28}}{double COLUMN_DESCRIPTION}
\item{\code{item29}}{double COLUMN_DESCRIPTION}
\item{\code{O1}}{double COLUMN_DESCRIPTION}
\item{\code{O2}}{double COLUMN_DESCRIPTION}
\item{\code{O3}}{double COLUMN_DESCRIPTION}
\item{\code{O4}}{double COLUMN_DESCRIPTION}
\item{\code{O5}}{double COLUMN_DESCRIPTION}
\item{\code{O6}}{double COLUMN_DESCRIPTION}
\item{\code{O7}}{double COLUMN_DESCRIPTION}
\item{\code{O8}}{double COLUMN_DESCRIPTION}
\item{\code{O9}}{double COLUMN_DESCRIPTION}
\item{\code{O10}}{double COLUMN_DESCRIPTION}
\item{\code{Total_Opennes}}{double COLUMN_DESCRIPTION}
\item{\code{E1}}{double COLUMN_DESCRIPTION}
\item{\code{E2}}{double COLUMN_DESCRIPTION}
\item{\code{E3}}{double COLUMN_DESCRIPTION}
\item{\code{E4}}{double COLUMN_DESCRIPTION}
\item{\code{E5}}{double COLUMN_DESCRIPTION}
\item{\code{E6}}{double COLUMN_DESCRIPTION}
\item{\code{E7}}{double COLUMN_DESCRIPTION}
\item{\code{E8}}{double COLUMN_DESCRIPTION}
\item{\code{Total_Extro}}{double COLUMN_DESCRIPTION}
\item{\code{N1}}{double COLUMN_DESCRIPTION}
\item{\code{N2}}{double COLUMN_DESCRIPTION}
\item{\code{N3}}{double COLUMN_DESCRIPTION}
\item{\code{N4}}{double COLUMN_DESCRIPTION}
\item{\code{N5}}{double COLUMN_DESCRIPTION}
\item{\code{N6}}{double COLUMN_DESCRIPTION}
\item{\code{N7}}{double COLUMN_DESCRIPTION}
\item{\code{N8}}{double COLUMN_DESCRIPTION}
\item{\code{Total_Neuro}}{double COLUMN_DESCRIPTION}
\item{\code{DIR1}}{double COLUMN_DESCRIPTION}
\item{\code{DIR2}}{double COLUMN_DESCRIPTION}
\item{\code{DI3}}{double COLUMN_DESCRIPTION}
\item{\code{DIR4}}{double COLUMN_DESCRIPTION}
\item{\code{DI5}}{double COLUMN_DESCRIPTION}
\item{\code{DIR6}}{double COLUMN_DESCRIPTION}
\item{\code{DI7}}{double COLUMN_DESCRIPTION}
\item{\code{DIR8}}{double COLUMN_DESCRIPTION}
\item{\code{DI9}}{double COLUMN_DESCRIPTION}
\item{\code{DI10}}{double COLUMN_DESCRIPTION}
\item{\code{DIR11}}{double COLUMN_DESCRIPTION}
\item{\code{DI12}}{double COLUMN_DESCRIPTION}
\item{\code{DI13}}{double COLUMN_DESCRIPTION}
\item{\code{DIR14}}{double COLUMN_DESCRIPTION}
\item{\code{DI15}}{double COLUMN_DESCRIPTION}
\item{\code{DI16}}{double COLUMN_DESCRIPTION}
\item{\code{DIR17}}{double COLUMN_DESCRIPTION}
\item{\code{DI18}}{double COLUMN_DESCRIPTION}
\item{\code{DIR19}}{double COLUMN_DESCRIPTION}
\item{\code{DI20}}{double COLUMN_DESCRIPTION}
\item{\code{DI21}}{double COLUMN_DESCRIPTION}
\item{\code{DIR22}}{double COLUMN_DESCRIPTION}
\item{\code{DIR23}}{double COLUMN_DESCRIPTION}
\item{\code{DIR24}}{double COLUMN_DESCRIPTION}
\item{\code{DI25}}{double COLUMN_DESCRIPTION}
\item{\code{DIR26}}{double COLUMN_DESCRIPTION}
\item{\code{DIR27}}{double COLUMN_DESCRIPTION}
\item{\code{DI28}}{double COLUMN_DESCRIPTION}
\item{\code{DI_Total}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{https://github.com/masiraji/tabledown/tree/main/data-raw}
}
\usage{
Rotter
}
\description{
This is the validation data of Bangla Rotter's Internal and External Scale.
}
\keyword{datasets}
