% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Spot}
\alias{Spot}
\title{Spot Data}
\format{
A data frame with 29 rows and 3 variables:
\describe{
\item{\code{activity}}{character Activity}
\item{\code{spot_type}}{character Progress Status}
\item{\code{spot_date}}{character Date of Reporting Progress}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
Spot
}
\description{
Additional demo data for GanTT
}
\keyword{datasets}
