% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.tab.R
\name{cfa.tab}
\alias{cfa.tab}
\title{A Function for Creating  Publication Quality Tables with CFA fit indices.}
\usage{
cfa.tab(x, robust = FALSE)
}
\arguments{
\item{x}{A lavaan class object.}

\item{robust}{If TRUE, will provide robust fit indices when applicable instead of the default indices.}
}
\description{
This function will create publication worthy tables with CFA fit indices from lavaan class object.
}
