% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des.tab.R
\name{des.tab}
\alias{des.tab}
\title{A Function for Descriptive data for item analysis}
\usage{
des.tab(df)
}
\arguments{
\item{df}{A data frame
table <- des.tab(data)}
}
\description{
This function will create a publication ready essential descriptive table for item analysis
}
\examples{
data <- BAAP::Rotter[, 11:31]
}
