% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggitem_info.R
\name{ggiteminfo}
\alias{ggiteminfo}
\title{A Function for Creating  Publication Quality IRT Item Characteristic Plot}
\usage{
ggiteminfo(model, item, theta)
}
\arguments{
\item{model}{A mirt package fitted object}

\item{item}{Item number (i.e. 1,2,3,4)}

\item{theta}{Theta range. Put only one number. Theta =3 will be considered as theta range (-3 to 3)}
}
\value{
A publication quality item information plot
}
\description{
This function will create publication worthy IRT item characteristic plot using ggplot2 from objects created from
mirt pack.
Using ggplot2 will enable the user to modify the tem characteristic plot in many ways
}
\examples{
data <- tabledown::Rotter[, 11:31]
model <- mirt::mirt(data, model = 1, itemtype = '2PL')

plot <- ggiteminfo(model, 1, 3)
}
