% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtestinfo.R
\name{ggtestinfo}
\alias{ggtestinfo}
\title{A Function for Creating  Publication Quality Item Response Theory based test information plot.}
\usage{
ggtestinfo(dataframe, model)
}
\arguments{
\item{dataframe}{your data.}

\item{model}{A mirt package fitted object.}
}
\value{
A publication quality Test information plot. Output object is a ggplot object.
}
\description{
This function will create publication worthy Item Response Theory based Test information plot using ggplot2 from objects created from
mirt pack.
Using ggplot2 will enable the user to modify the Item plot.
}
\examples{
data <- tabledown::Rotter[, 11:31]
model <- mirt::mirt(data, model = 1, itemtype = '2PL')

plot <- ggtestinfo(data, model)
}
